requirements.txt
youtube-dl==2021.12.17
python-dotenv==0.21.1
pyTelegramBotAPI==4.7.0